+++ src/index.js
@@ -1,0 +1,3 @@
+import StarRating from './star-rating.vue' 
+
+export default StarRating