+++ webpack.config.js
@@ -1,0 +1,5 @@
+function buildConfig(env) {
+  return require('./build/' + env + '.js');
+}
+
+module.exports = buildConfig;