+++ changelog.md
@@ -2,6 +2,13 @@
 - Changes before version 1.2.1 not documented (see commit history)
 - Project follows [Semantic Versioning](http://semver.org/)
 
+## [Unreleased]
+
+- Webpack added and is now being used for build and development
+- ESLint added
+- RTL Support Added
+- fixed-points prop added
+
 ## Version 1.3.0
 
 ### Added